backend/recommendations/requirements.txt
flask
flask-cors
python-dotenv
supabase
pyjwt
python-dateutil
